requirements.txt
pyqt>=5.9.2
requests>=2.22.0
python-dotenv>=0.19.0